backend/pyproject.toml
[tool.poetry]
name = "django-vue-template"
version = "0.1.0"
description = ""
authors = ["MarcGroc <marcgroc12@gmail.com>"]
readme = "README.md"
packages = [{include = "backend"}]
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Basic dependencies used by all environments
[tool.poetry.dependencies]
python = "^3.11"
celery = {extras = ["redis"], version = "^5.3.6"}
flower = "^2.0.1"
redis = "^5.0.1"
drf-spectacular = "^0.27.1"

#Django dependencies
django = "4.2.10"
djangorestframework = "^3.14.0"
environ = "^1.0"
django-environ = "^0.11.2"
django-celery-beat = "2.5.0"
django-allauth = "0.61.0"
django-redis = "^5.4.0"
django-axes = {extras = ["ipware"], version = "^6.3.0"}
flake8 = "^7.0.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
gevent = "^24.2.1"
watchfiles = "^0.21.0"
argon2-cffi = "^23.1.0"
loguru = "^0.7.2"
django-prometheus = "^2.3.1"
dj-rest-auth = "^5.1.0"
stripe = "^9.8.0"
django-cors-headers = "^4.3.1"


[tool.poetry.group.dev.dependencies]
# Development dependencies for development and testing
pytest = "^8.0.0"
django-extensions = "^3.2.3"
mypy = "^1.8.0"
pytest-sugar = "^1.0.0"
psycopg2 = {extras = ["c"], version = "^2.9.9"}
psycopg2-binary = "^2.9.9"
flake8-isort = "^6.1.1"
pre-commit = "^3.6.1"
pytest-django = "^4.8.0"
factory-boy = "^3.3.0"
ruff = "^0.2.1"
tox = "^4.12.1"
django-debug-toolbar = "^4.3.0"


[tool.poetry.group.prod.dependencies]
# Production dependencies
django-storages = "^1.14.2"


# Tools configuration
[tool.pytest.ini_options]
addopts = "--ds=config.settings.test --reuse-db"
python_files = "test_*.py"


[tool.ruff]
line-length = 119
exclude = ["migrations", "tests", ".git", ".pytest_cache",".ruff_cache",
    ".venv", "venv"]

[tool.ruff.lint]
select = ["E", "F"]
ignore = ["F405"]


[tool.isort]
profile = "django"
line_length = 119

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py38, py39, py310, py311, py312

[testenv]
changedir = tests
deps = pytest
commands = pytest
"""
